Code cleanup, split into more files for clarity
